+++ packages/expo/types/global.d.ts
@@ -29,3 +29,49 @@ declare module '*.module.scss' {
 declare module '*.css';
 declare module '*.sass';
 declare module '*.scss';
+
+// Allow for image asset imports
+declare module '*.png' {
+  const path: string;
+  export default path;
+}
+
+declare module '*.jpg' {
+  const path: string;
+  export default path;
+}
+
+declare module '*.jpeg' {
+  const path: string;
+  export default path;
+}
+
+declare module '*.webp' {
+  const path: string;
+  export default path;
+}
+
+declare module '*.avif' {
+  const path: string;
+  export default path;
+}
+
+declare module '*.gif' {
+  const path: string;
+  export default path;
+}
+
+declare module '*.ico' {
+  const path: string;
+  export default path;
+}
+
+declare module '*.bmp' {
+  const path: string;
+  export default path;
+}
+
+declare module '*.svg' {
+  const path: string;
+  export default path;
+}